requirements.txt
yt-dlp>=2025.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.3
openai>=1.40.0
pytest>=8.3.0
